src/scraper/requirements.txt
# src/scraper/requirements.txt
Scrapy
psycopg2-binary
python-dotenv

# --- ORM y Migraciones ---
SQLAlchemy
alembic